refactor: introduce interface mailfilter.Trx and a testing implementation testtrx.Trx
With this new testtrx.Trx users of the mailfilter milter abstraction can easily write unit tests for their decision function. before this refactoring the decision function could not easily be unit tested because generating mailfilter.Transaction structs was not possible outside of this library.

BREAKING CHANGE: the mailfilter signatures changed significantly because we now pass a Trx interface to the decision function instead of a Transaction struct. The changes library users need to do are mostly calling methods instead of accessing public fields of the transaction.