feat: drop Trx.Log and WithSyslog from mailfilter
Trx.Log would just add the QueueId to your log string, that's not that much of a convenience to warrant the tight coupling with the log standard library. The implementation cannot be used with structured logging libraries so wen users of this library use a more sophisticated logging they cannot use this little helper anyway.

BREAKING CHANGE: you should bring your own logging