feat: add HeadersEnforceOrder() to Trx
Calling HeadersEnforceOrder on the Trx will activate a special workaround for Sendmail to enforce our header ordering.

The workaround is to unconditionally  delete all existing headers and add (the possibly changed) headers back. This is a significant overhead, so you should only enable this workaround when you really need to enforce your header order.

This commit also changes the semantics of InsertBefore* method calls when called for the very first header. Before this the header would have been inserted in-front of the `Received` header of the MTA (position 0) and not in-front of the first header (position 1).