refactor: update Milter interface & introduce Modifier interface
BREAKING CHANGE: One Milter instance will now handle one milter connection. It must be able to handle multiple messages. To facilitate this, a new method NewConnection was added to the Milter interface.

BREAKING CHANGE: The struct Modifier was changed to an interface Modifier.
All Milter implementations need to use `m milter.Modifier` instead of `m *milter.Modifier` in their arguments.
This improves the testability of Milter implementations since you can easily create mock Modifier implementations that can be passed in to your own Milter implementations (that was not possible before)

BREAKING CHANGE: the default read timeout for Server changed from 10 seconds to 0 seconds. The 10 seconds default was actually not used - we always used 0 (no timeout) before, so this should not really be a breaking change.